requirements.txt
PyMuPDF==1.24.14
Pillow==11.0.0
playwright==1.49.0
